refactoring read/write to move logic inside volume layer
