+++ src/lib/CMakeLists.txt
@@ -7,7 +7,6 @@ target_sources(${PROJECT_NAME}_core PRIVATE
     homeblks_impl.cpp
     volume_mgr.cpp
     listener.cpp
-    index.cpp
     common.cpp
 )
 target_link_libraries(${PROJECT_NAME}_core
