+++ src/lib/volume_mgr.cpp
@@ -20,14 +20,6 @@
 #include "homeblks_impl.hpp"
 
 namespace homeblocks {
-
-static VolumeError to_volume_error(std::error_code ec) {
-    switch (ec.value()) {
-    default:
-        return VolumeError::UNKNOWN;
-    }
-}
-
 std::shared_ptr< VolumeManager > HomeBlocksImpl::volume_manager() { return shared_from_this(); }
 
 void HomeBlocksImpl::on_vol_meta_blk_found(sisl::byte_view const& buf, void* cookie) {
@@ -142,200 +134,11 @@ bool HomeBlocksImpl::get_stats(volume_id_t id, VolumeStats& stats) const { retur
 void HomeBlocksImpl::get_volume_ids(std::vector< volume_id_t >& vol_ids) const {}
 
 VolumeManager::NullAsyncResult HomeBlocksImpl::write(const VolumePtr& vol_ptr, const vol_interface_req_ptr& vol_req) {
-
-    // Step 1. Allocate new blkids. Homestore might return multiple blkid's pointing
-    // to different contigious memory locations.
-    auto data_size = vol_req->nlbas * vol_ptr->rd()->get_blk_size();
-    std::vector< homestore::MultiBlkId > new_blkids;
-    auto result = vol_ptr->rd()->alloc_blks(data_size, homestore::blk_alloc_hints{}, new_blkids);
-    if (result) {
-        LOGE("Failed to allocate blocks");
-        return folly::makeUnexpected(VolumeError::NO_SPACE_LEFT);
-    }
-
-    // Step 2. Write the data to those allocated blkids.
-    sisl::sg_list data_sgs;
-    data_sgs.iovs.emplace_back(iovec{.iov_base = vol_req->buffer, .iov_len = data_size});
-    data_sgs.size = data_size;
-    return vol_ptr->rd()
-        ->async_write(new_blkids, data_sgs, vol_req->part_of_batch)
-        .thenValue([this, vol_ptr, vol_req,
-                    new_blkids = std::move(new_blkids)](auto&& result) -> VolumeManager::NullAsyncResult {
-            if (result) { return folly::makeUnexpected(VolumeError::DRIVE_WRITE_ERROR); }
-
-            using homestore::BlkId;
-            std::vector< BlkId > old_blkids;
-            std::unordered_map< lba_t, BlockInfo > blocks_info;
-            auto blk_size = vol_ptr->rd()->get_blk_size();
-            auto data_size = vol_req->nlbas * blk_size;
-            auto data_buffer = vol_req->buffer;
-            lba_t start_lba = vol_req->lba;
-            for (auto& blkid : new_blkids) {
-                DEBUG_ASSERT_EQ(blkid.num_pieces(), 1, "Multiple blkid pieces");
-
-                // Split the large blkid to individual blkid having only one block because each LBA points
-                // to a blkid containing single blk which is stored in index value. Calculate the checksum for each
-                // block which is also stored in index.
-                for (uint32_t i = 0; i < blkid.blk_count(); i++) {
-                    auto new_bid = BlkId{blkid.blk_num() + i, 1 /* nblks */, blkid.chunk_num()};
-                    auto csum = crc16_t10dif(init_crc_16, static_cast< unsigned char* >(data_buffer), blk_size);
-                    blocks_info.emplace(start_lba + i, BlockInfo{new_bid, BlkId{}, csum});
-                    data_buffer += blk_size;
-                }
-
-                // Step 3. For range [start_lba, end_lba] in this blkid, write the values to index.
-                // Should there be any overwritten on existing lbas, old blocks to be freed will be collected
-                // in blocks_info after write_to_index
-                lba_t end_lba = start_lba + blkid.blk_count() - 1;
-                auto status = write_to_index(vol_ptr, start_lba, end_lba, blocks_info);
-                if (!status) { return folly::makeUnexpected(VolumeError::INDEX_ERROR); }
-
-                start_lba = end_lba + 1;
-            }
-
-            // Collect all old blocks to write to journal.
-            for (auto& [_, info] : blocks_info) {
-                if (info.old_blkid.is_valid()) { old_blkids.emplace_back(info.old_blkid); }
-            }
-
-            auto csum_size = sizeof(homestore::csum_t) * vol_req->nlbas;
-            auto old_blkids_size = sizeof(BlkId) * old_blkids.size();
-            auto key_size = sizeof(VolJournalEntry) + csum_size + old_blkids_size;
-
-            auto req = repl_result_ctx< VolumeManager::NullResult >::make(
-                sizeof(HomeBlksMessageHeader) /* header size */, key_size);
-            req->vol_ptr_ = vol_ptr;
-            req->header()->msg_type = HomeBlksMsgType::WRITE;
-            // Store volume id for recovery path (log replay)
-            req->header()->volume_id = vol_ptr->id();
-
-            // Step 4. Store lba, nlbas, list of checksum of each blk, list of old blkids as key in the journal.
-            // New blkid's are written to journal by the homestore async_write_journal. After journal flush, on_commit
-            // will be called where we free the old blkid's and the write iscompleted.
-            VolJournalEntry hb_key{vol_req->lba, vol_req->nlbas, static_cast< uint16_t >(old_blkids.size())};
-            auto key_buf = req->key_buf().bytes();
-            std::memcpy(key_buf, &hb_key, sizeof(VolJournalEntry));
-            key_buf += sizeof(VolJournalEntry);
-
-            auto lba = vol_req->lba;
-            for (lba_count_t count = 0; count < vol_req->nlbas; count++) {
-                std::memcpy(key_buf, &blocks_info[lba].checksum, sizeof(homestore::csum_t));
-                key_buf += sizeof(homestore::csum_t);
-                lba++;
-            }
-
-            for (auto& blkid : old_blkids) {
-                std::memcpy(key_buf, &blkid, sizeof(BlkId));
-                key_buf += sizeof(BlkId);
-            }
-
-            vol_ptr->rd()->async_write_journal(new_blkids, req->cheader_buf(), req->ckey_buf(), data_size, req);
-            return req->result().deferValue([this](const auto&& result) -> folly::Expected< folly::Unit, VolumeError > {
-                if (result.hasError()) {
-                    auto err = result.error();
-                    return folly::makeUnexpected(err);
-                }
-                return folly::Unit();
-            });
-        });
-}
-
-void HomeBlocksImpl::generate_blkids_to_read(const index_kv_list_t& index_kvs, read_blks_list_t& blks_to_read) {
-    for(uint32_t i = 0, start_idx = 0; i < index_kvs.size(); ++i) {
-        auto const& [key, value] = index_kvs[i];
-        bool is_contiguous = (i == 0 || (value.blkid().blk_num() == index_kvs[i-1].second.blkid().blk_num() + 1
-                                && value.blkid().chunk_num() == index_kvs[i-1].second.blkid().chunk_num()));
-        if(is_contiguous && i < index_kvs.size() - 1) {
-            // continue to the next entry if it is contiguous
-            continue;
-        }
-        // prepare the previous contiguous blkids to read
-        auto blk_num = index_kvs[start_idx].second.blkid().blk_num();
-        auto chunk_num = index_kvs[start_idx].second.blkid().chunk_num();
-        // if the last entry is part of the contiguous block,
-        // we need to account for it in the blk_count
-        auto blk_count = is_contiguous ? (i - start_idx + 1) : (i - start_idx);
-        blks_to_read.emplace_back(index_kvs[start_idx].first.key(), homestore::MultiBlkId(blk_num, blk_count, chunk_num));
-        start_idx = i;
-        if(!is_contiguous && i == index_kvs.size() - 1) {
-            // if the last entry is not contiguous, we need to add it as well
-            blks_to_read.emplace_back(key.key(), homestore::MultiBlkId(value.blkid().blk_num(), 1, value.blkid().chunk_num()));
-        }
-    }
-}
-
-void HomeBlocksImpl::submit_read_to_backend(read_blks_list_t const& blks_to_read, const vol_interface_req_ptr& req, 
-                                            const VolumePtr& vol, std::vector< folly::Future< std::error_code > >& futs) {
-    auto* read_buf = req->buffer;
-    DEBUG_ASSERT(read_buf != nullptr, "Read buffer is null");
-    for(uint32_t i = 0, prev_lba = req->lba, prev_nblks = 0; i < blks_to_read.size(); ++i) {
-        auto const& [start_lba, blkids] = blks_to_read[i];
-        DEBUG_ASSERT(start_lba >= prev_lba + prev_nblks, "Invalid start lba: {}, prev_lba: {}, prev_nblks: {}",
-                     start_lba, prev_lba, prev_nblks);
-        auto holes_nblks = start_lba - (prev_lba + prev_nblks);
-        read_buf += (holes_nblks * vol->rd()->get_blk_size());
-        sisl::sg_list sgs;
-        sgs.size = blkids.blk_count() * vol->rd()->get_blk_size();
-        sgs.iovs.emplace_back(iovec{.iov_base = read_buf, .iov_len = sgs.size});
-        read_buf += sgs.size;
-        futs.emplace_back(vol->rd()->async_read(blkids, sgs, sgs.size, req->part_of_batch));
-        prev_lba = start_lba;
-        prev_nblks = blkids.blk_count();
-    }
+    return vol_ptr->write(vol_req);
 }
 
 VolumeManager::NullAsyncResult HomeBlocksImpl::read(const VolumePtr& vol, const vol_interface_req_ptr& req) {
-    // TODO: check if the system is accepting ios (shutdown in progress etc)
-    RELEASE_ASSERT(vol != nullptr, "VolumePtr is null");
-    // Step 1: get the blk ids from index table
-    vol_read_ctx read_ctx{.buf = req->buffer, .start_lba = req->lba, .blk_size = vol->rd()->get_blk_size()};
-    if(auto index_resp = read_from_index(vol, req, read_ctx.index_kvs); index_resp.hasError()) {
-        LOGE("Failed to read from index table for range=[{}, {}], volume id: {}, error: {}",
-             req->lba, req->end_lba(), boost::uuids::to_string(vol->id()), index_resp.error());
-        return index_resp;
-    }
-    if (read_ctx.index_kvs.empty()) {
-        return folly::Unit();
-    }
-    
-    // Step 2: Consolidate the blocks by merging the contiguous blkids
-    std::vector< folly::Future< std::error_code > > futs;
-    read_blks_list_t blks_to_read;
-    generate_blkids_to_read(read_ctx.index_kvs, blks_to_read);
-
-    // Step 3: Submit the read requests to backend
-    submit_read_to_backend(blks_to_read, req, vol, futs);
-
-    // Step 4: verify the checksum after all the reads are done
-    return  folly::collectAllUnsafe(futs).thenValue([this, read_ctx = std::move(read_ctx)](auto&& vf) -> VolumeManager::Result< folly::Unit > {
-        for (auto const& err_c : vf) {
-            if (sisl_unlikely(err_c.value())) {
-                auto ec = err_c.value();
-                return folly::makeUnexpected(to_volume_error(ec));
-            }
-        }
-        // verify the checksum and return
-        return verify_checksum(read_ctx);
-    });
-}
-
-VolumeManager::Result< folly::Unit > HomeBlocksImpl::verify_checksum(vol_read_ctx const& read_ctx) {
-    auto read_buf = read_ctx.buf;
-    for(uint64_t cur_lba = read_ctx.start_lba, i = 0; i < read_ctx.index_kvs.size(); ++i, ++cur_lba) {
-        auto const& [key, value] = read_ctx.index_kvs[i];
-        // ignore the holes
-        if(cur_lba != key.key()) {
-            read_buf += (key.key() - cur_lba) * read_ctx.blk_size;
-            cur_lba = key.key();
-        }
-        auto checksum = crc16_t10dif(init_crc_16, static_cast< unsigned char* >(read_buf), read_ctx.blk_size);
-        if(checksum != value.checksum()) {
-            LOGE("crc mismatch for lba: {}, blk id {}, expected: {}, actual: {}", cur_lba, value.blkid().to_string(), value.checksum(), checksum);
-            return folly::makeUnexpected(VolumeError::CRC_MISMATCH);
-        }
-        read_buf += read_ctx.blk_size;
-    }
-    return folly::Unit();
+    return vol->read(req);
 }
 
 VolumeManager::NullAsyncResult HomeBlocksImpl::unmap(const VolumePtr& vol, const vol_interface_req_ptr& req) {
@@ -350,7 +153,7 @@ void HomeBlocksImpl::on_write(int64_t lsn, const sisl::blob& header, const sisl:
                               cintrusive< homestore::repl_req_ctx >& ctx) {
     repl_result_ctx< VolumeManager::NullResult >* repl_ctx{nullptr};
     if (ctx) { repl_ctx = boost::static_pointer_cast< repl_result_ctx< VolumeManager::NullResult > >(ctx).get(); }
-    auto msg_header = r_cast< HomeBlksMessageHeader* >(const_cast< uint8_t* >(header.cbytes()));
+    auto msg_header = r_cast< MsgHeader* >(const_cast< uint8_t* >(header.cbytes()));
 
     // Key contains the list of checksums and old blkids. Before we ack the client
     // request, we free the old blkid's. Also if its recovery we overwrite the index
@@ -381,7 +184,7 @@ void HomeBlocksImpl::on_write(int64_t lsn, const sisl::blob& header, const sisl:
             // We ignore the existing values we got in blocks_info from index as it will be
             // same checksum, blkid we see in the journal entry.
             lba_t end_lba = start_lba + blkid.blk_count() - 1;
-            auto status = write_to_index(vol_ptr, start_lba, end_lba, blocks_info);
+            auto status = vol_ptr->write_to_index(start_lba, end_lba, blocks_info);
             RELEASE_ASSERT(status, "Index error during recovery");
             start_lba = end_lba + 1;
         }
